Removing gather of pod logs
We will work in other PR/card adding relationship of objects managed
by cluster operator to RelatedObjects, just so must-gather binary
pulls the pod logs for free and there's no need to pull that in bash
script.
We leave therefore in the bash scripts the gathering of iptables,
ovs-vsctl dump and vswitchd/ovsdb-server logs (this is for better
visualization, since the OVS pods have those logs interleaved).