Add collection-script for sosreports
This allows users to gather sosreports on all or on a subset of
nodes of their cluster. This follows the same idea as the
gather_core_dumps scripts. This is not run by default, instead users
must explicitly run this with /usr/bin/gather-sosreports
and they can provide a label selector and overwrite for image
and registry, e.g.:
    -l kubernetes.io/hostname=openshift-worker-0
    -r registry.example.com:5000
    -i support-tools:0.0.2

Signed-off-by: Andreas Karis <ak.karis@gmail.com>