Fix the metallb operator namespace filter
Currently, the gather_metallb_logs script is extracting the operator
namespace by selecting the subscription with the metadata.name equal
to "metallb-operator". However, it would be more correct to check
against the spec.name that points to the name of the operator,
as the value of this field is constant. unlike the other that can
be called "metallb-operator-sub".

Note that I didn't use `--field-selector` in the new implementation
as spec.name field label is not supported.