Gather MCO's on-disk configs from degraded nodes
There have been a lot of bugs recently where the MCO doesn't render the
same thing during bootstrap and live operation, usually due to race
conditions introduced elsewhere. In order to chase these we often have
to ask reporters/customers to manually collect on-disk files from
the node. (example: https://issues.redhat.com/browse/OCPBUGS-12741)

This adds a gather script that collects the on-disk
mcs-machine-config-content.json and the currentconfig from
any node the MCO has marked degraded -- so in the event that the
degradation is caused by a rendering/mismatch/validation issue we can
address it without having to ask for further manual collection. This
will also make sure that if we hit these issues in CI we can more
easily troubleshoot them.