Assert that only one subscription exists
Users that mistakenly misconfigure and create two instances (or more)
of the subscription resource causes some gather scripts to fail.

For example, with 2 different sriov subscriptions, the SRIOV_NS will equal
"openshift-sriov-network-operatoropenshift-sriov-network-operator"
which is invalid.

Here we add "{{\n}}" and with that, check whether we found more than
one subscription, and if so, exit with an error.

This fix is applied to all of the gather scripts that use this method:
ingress_node_firewall, metallb, network_observability, nmstate and sriov

Signed-off-by: liornoy <lnoy@redhat.com>
Co-authored-by: Vladislav Walek <22072258+vwalek@users.noreply.github.com>