Use oc get daemonset to identify the NTO image
The previous method of using oc adm release info has been
problematic for CI and restricted network environment clusters.

The issue was always the authentication to the registry
and access to the release image from the inside of must-gather.

Also the oc adm release info tool runs inside a container and
does not respect the node level ImageContentSourcePolicy.

For all these reasons I want to retrieve the image name
from the configured NTO deployment instead.

Signed-off-by: Martin Sivak <msivak@redhat.com>