OPNET-470: Collect host networking logs
It happens very often that must-gather is the only bundle provided to
the networking team when asking to help debugging various issues. In
reality, in majority of those cases we would rather get sosreport and
not must-gather because the former contains useful outputs from `ip`
command.

With this PR we are trying to get the most basic networking informations
already into must-gather to speed up debugging.

Contributes-to: OPNET-470
Contributes-to: OCPBUGS-26217
Contributes-to: OCPBUGS-29624