OCPBUGS-34360: Run ppc node collection in parallel
The PPC node data collection was running in serial and that
increased the time needed to collect the whole cluster.
This change executes each per-node sequence in parallel and
waits for all the sub tasks.

The only disadvantage is having a per-node process (potentially
hundreds in huge clusters), however this case has always been
in the code as well. The load on the controlling must-gather pod
should be low as it only executes remote commands and waits for
results.

This can be tested (at least while the PR is open) by executing:

oc adm must-gather --image quay.io/msivak/must-gather:test416c