gather_ppc: fallback option when NTO image not found
gather_ppc script is looking for NTO image, becuase it contains all the
tools that needed.

On Hypershift platfrom, the control-plane components are located on a different cluster, so
the script can't find NTO image since the controller is running on the
remote (management) cluster.

This commit suggests to add a fallback method:
In case NTO deployment was not found, the script tries
to get NTO image name from the release image.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>