Fix collecting IPsec data for upgrade
After IPsec is upgraded from 4.14 into later versions, API option for IPsec in
networks.operator.openshift.io cluster object continues to have {} in the ipsecConfig
and may not be migrated to newer version containing IPsec mode option. This is supported
in OCP for backward compatibility, so mg needs fixing to collect ipsec data when API
configured with old option.

Signed-off-by: Periyasamy Palanisamy <pepalani@redhat.com>