gather_ppc: detect NTO image from TuneD pod on HCP
When running the MG on hosted cluster, NTO deployment is running on the management cluster, so MG cannot access NTO image.

Previously we added https://github.com/openshift/must-gather/pull/436  as a fallback method for such cases but there are customer setups with no access to the release image.

Hence, we are adding another method to fetch NTO image directly from TuneD pods.
This method will be exercised only when running MG on the hosted cluster.

Signed-off-by: Talor Itzhak <titzhak@redhat.com>